Move worker_proto defs out of remote-store.cc to own file
